requirements.txt
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
pip-chill==1.0.3
psycopg2-binary==2.9.9
